% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Extended Approach cNB ml.R
\name{ml.ex.cnb}
\alias{ml.ex.cnb}
\title{MLE of contaminated negative binomial regression model
extended to include covariate estimation of all distribution parameters.}
\usage{
ml.ex.cnb(
  formula,
  alpha.formula = ~1,
  delta.formula = ~1,
  eta.formula = ~1,
  data,
  start = NULL,
  method = "Nelder-Mead",
  reltol = 1e-15,
  maxit = 1000,
  hessian = T
)
}
\arguments{
\item{formula}{an object of class 'formula': a symbolic description of the model to be fitted.}

\item{alpha.formula}{A formula for the dispersion parameter (sigma). Defaults to ~1.}

\item{delta.formula}{A formula for the proportion of extreme values parameter (delta). Defaults to ~1.}

\item{eta.formula}{A formula for the inflation parameter (eta). Defaults to ~1.}

\item{data}{a mandatory data frame containing the variables in the model.}

\item{start}{vector of initial values. If NULL, then values produced by glm.nb are used as initial values}

\item{method}{optimization method to be used. Default is "Nelder-Mead". Other options include "CG" and "BFGS".}

\item{reltol}{relative convergence tolerance in each optimization process. Defaults to 1e-15.}

\item{maxit}{the number of inner iterations in each optimization process. Defaults to 10000.}

\item{em.tol}{the EM convergence tolerance. Defaults to 1e-10.}

\item{em.maxit}{the number of EM iterations. Defaults to 1000.}
}
\value{
A list of elements:
   \item{results}{A data frame with parameter estimates and standard errors. }
   \item{alpha}{Maximum likelihood estimate of alpha.}
   \item{delta}{Maximum likelihood estimate of delta.}
   \item{beta}{Maximum likelihood estimates of the regression coefficients.}
   \item{mu}{Maximum likelihood estimates of the mean parameter.}
   \item{X}{The design matrix for mu.}
   \item{U}{The design matrix for alpha.}
   \item{V}{The design matrix for delta.}
   \item{Z}{The design matrix for eta.}
   \item{y}{The response variable.}
   \item{loglike}{The log-likelihood value at convergence.}
   \item{AIC}{Akaike Information Criterion (AIC) for the fitted model.}
   \item{BIC}{Bayesian Information Criterion (BIC) for the fitted model.}
   \item{HQIC}{Hannan-Quinn Information Criterion (HQIC) for the fitted model.}
}
\description{
MLE of the contaminated negative binomial regression model via an Expectation-Maximization algorithm.
}
\details{
The \code{ml.cnb} function fits the contaminated negative binomial regression model (https://github.com/arnootto/cNB/tree/master).
}
\examples{
library(AER)
library(cNB)

# Load NMES1988 dataset
data("NMES1988")
head(NMES1988)

# Construct design matrix
x <- model.matrix(~ visits + nvisits + ovisits + novisits + emergency + hospital +
                  health + chronic + adl + region + afam + gender + married +
                  school + income + employed + insurance + medicaid,
                  data = NMES1988)
x <- as.data.frame(x)

options(contrasts = c("contr.treatment", "contr.poly"))

# Regression on mean, dispersion, delta, and eta
cnb_ex_trips_mu_alpha_delta_eta <- ml.ex.cnb(
  formula = visits ~ hospital + insuranceyes + healthpoor + chronic + adllimited +
                      gendermale + school + afamyes,
  alpha.formula = ~ adllimited,
  delta.formula = ~ nvisits,
  eta.formula = ~ chronic,
  data = x,
  method = "BFGS"
)

# Parameter estimates
round(cnb_ex_trips_mu_alpha_delta_eta$results, 4)

# Information criteria and log-likelihood
round(cnb_ex_trips_mu_alpha_delta_eta$AIC, 4)
round(cnb_ex_trips_mu_alpha_delta_eta$BIC, 4)
round(cnb_ex_trips_mu_alpha_delta_eta$HQIC, 4)
round(cnb_ex_trips_mu_alpha_delta_eta$loglike, 4)
}
